+++ src/fhir/patient-helper.ts
@@ -2,15 +2,16 @@ import { SearchParams } from "fhir-kit-client";
 import { getIncludedResources } from "./bundle";
 import { PatientModel } from "./models";
 import { searchBuilderRecords } from "./search-helpers";
-import { SYSTEM_ZUS_UPI_RECORD_TYPE } from "./system-urls";
+import { SYSTEM_ZUS_OWNER, SYSTEM_ZUS_UPI_RECORD_TYPE } from "./system-urls";
+import { useQueryWithPatient } from "..";
+import { CTWRequestContext } from "../components/core/providers/ctw-context";
+import { useQueryWithCTW } from "../components/core/providers/use-query-with-ctw";
 import {
   createGraphqlClient,
-  CTWRequestContext,
   fqsRequest,
   GraphqlPageInfo,
   MAX_OBJECTS_PER_REQUEST,
-  useQueryWithCTW,
-} from "..";
+} from "../services/fqs/client";
 import {
   PatientGraphqlResponse,
   patientsForBuilderQuery,
@@ -18,8 +19,32 @@ import {
 } from "@/services/fqs/queries/patients";
 import { errorResponse } from "@/utils/errors";
 import { pickBy } from "@/utils/nodash";
-import { QUERY_KEY_PATIENTS_LIST_FQS, QUERY_KEY_PATIENTS_LIST_ODS } from "@/utils/query-keys";
+import {
+  QUERY_KEY_MATCHED_PATIENTS,
+  QUERY_KEY_PATIENTS_LIST_FQS,
+  QUERY_KEY_PATIENTS_LIST_ODS,
+} from "@/utils/query-keys";
+import { sort } from "@/utils/sort";
+import { withTimerMetric } from "@/utils/telemetry";
 import { hasNumber } from "@/utils/types";
+
+export function useMatchedPatients() {
+  const matchedPatientsQuery = useQueryWithPatient(
+    QUERY_KEY_MATCHED_PATIENTS,
+    [],
+    withTimerMetric(
+      async (requestContext: CTWRequestContext, patient: PatientModel) =>
+        getPatientsForUPIDFQS(requestContext, patient, {
+          tag: {
+            nonematch: [`${SYSTEM_ZUS_OWNER}|builder/${requestContext.builderId}`],
+          },
+        }),
+      "req.timing.matched_patients"
+    )
+  );
+
+  return matchedPatientsQuery;
+}
 
 export async function getPatientsForUPIDFQS(
   requestContext: CTWRequestContext,
@@ -50,7 +75,9 @@ export async function getPatientsForUPIDFQS(
   }
 }
 
-export async function getBuilderFhirPatient(
+// Returns a single FHIR patient given a patientID and systemURL.
+// If multiple patients are found then it returns the last updated.
+export async function getBuilderFhirPatientByExternalIdentifier(
   requestContext: CTWRequestContext,
   patientID: string,
   systemURL: string,
@@ -77,7 +104,22 @@ export async function getBuilderFhirPatient(
     );
   }
 
-  return new PatientModel(patients[0], getIncludedResources(bundle));
+  return new PatientModel(
+    sort(patients, "meta.lastUpdated", "desc", true)[0],
+    getIncludedResources(bundle)
+  );
+}
+
+export async function getPatientByID(
+  requestContext: CTWRequestContext,
+  patientID: string
+): Promise<PatientModel> {
+  const response = await requestContext.fhirClient.read({
+    resourceType: "Patient",
+    id: patientID,
+  });
+
+  return new PatientModel(response as fhir4.Patient);
 }
 
 export function usePatientsListFQS(pageSize: number, cursor: string) {